hw4-vm-http-client/requirements.txt

functions-framework==3.4.0
google-cloud-storage==2.11.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
waitress==2.1.2